requirements.txt
Django==2.2.6
django-ckeditor==5.7.1
django-contrib-comments==1.9.1
django-crispy-forms==1.7.2
django-forms-builder==0.4.7
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
docutils==0.15.2
mysqlclient==1.4.4
Pillow==6.2.0
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
yapf==0.28.0
